feat(prerequisite): add pids.max check and enhance UID/capability prerequisites
Add new prerequisite `pids.max` check under cgroups/pids to detect limited or unlimited PID limits,
including corresponding E2E test and configuration.
Extend capability checks with CAP_SYS_RESOURCE support.
Refactor user UID prerequisite to support ruid/euid/suid checking with flexible configuration.
Fix MaxUserNamespaces type for greater-than comparison.

+++ prerequisite/sysctl/user/max_user_namespaces.go
@@ -10,7 +10,7 @@ import (
 
 type MaxUserNamespaces struct {
 	prerequisite.BasePrerequisite
-	GreaterThan int
+	GreaterThan uint64
 }
 
 func (p *MaxUserNamespaces) Check() (bool, error) {
